gnrc IPv6: use uint16 for l2addr_len
netopt uses 16 bit integers for address option, IPv6 and 6LoWPAN need to reflect this.

+++ sys/include/net/gnrc/sixlowpan/nd.h
@@ -161,7 +161,7 @@ void gnrc_sixlowpan_nd_uc_rtr_sol(gnrc_ipv6_nc_t *nce);
  * @return  -ENOBUFS, if @p l2addr_len was smaller than the resulting @p l2addr
  *          would be long.
  */
-kernel_pid_t gnrc_sixlowpan_nd_next_hop_l2addr(uint8_t *l2addr, uint8_t *l2addr_len,
+kernel_pid_t gnrc_sixlowpan_nd_next_hop_l2addr(uint8_t *l2addr, uint16_t *l2addr_len,
                                                kernel_pid_t iface, ipv6_addr_t *dst);
 
 /**
